requirements.txt
streamlit>=1.29.0
opencv-python>=4.5.3
numpy>=1.21.0
Pillow>=8.4.0
